parser/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-pptx==1.0.2
PyMuPDF==1.25.1
python-multipart==0.0.20
